Open source release of Orange modifications
Clean up, complete open source publication.

Co-authored-by: Pierre-Yves Lapersonne <pierreyves.lapersonne@orange.com>
Co-authored-by: Pierre-Yves Lapersonne <dev@pylapersonne.info>

In 2018, the "Tool Factory" project of INNOV/TGI/IVA at Orange studied the Tapster project, but this project was quickly abandoned.
The only developer involved continued, in a personal context, to work on the project and enrich it.
In 2024, it was noted that it was planned to publish elements of this project in open source, this request having not been successful because the request management tool was closed, the open source referent in charge of the request at the time retired, and the main developer left the service.
This publication aims to correct this error; the main developer of the fork having agreed to cede his rights on this fork while keeping this fork.
Note that this situation arose due to a lack of means, training and acculturation to open source within the company.
The situation was resolved by a GitHub administrator, himself an open source referent, main developer of the said project, and who has since become a trainer to try to improve this grotesque and dramatic situation.

Side fork: https://github.com/pylapp/tapsterbot

Signed-off-by: Pierre-Yves Lapersonne <pierreyves.lapersonne@orange.com>
Signed-off-by: Pierre-Yves Lapersonne <dev@pylapersonne.info>